backend/requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
cachelib==0.13.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Django==5.1.7
djangorestframework==3.15.2
filelock==3.18.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Caching==2.3.1
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-MySQL==1.6.0
Flask-SQLAlchemy==3.1.1
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
h11==0.14.0
huggingface-hub==0.31.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
mpmath==1.3.0
mysqlclient==2.2.7
narwhals==1.35.0
networkx==3.4.2
numpy==2.2.5
packaging==24.2
pandas==2.2.3
pillow==11.2.1
protobuf==5.29.4
pyarrow==19.0.1
pydantic==2.11.3
pydantic_core==2.33.1
pydeck==0.9.1
PyJWT==2.10.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.24.0
safetensors==0.5.3
setuptools==80.7.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.39
sqlparse==0.5.3
starlette==0.46.2
streamlit==1.44.1
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.21.1
toml==0.10.2
torch==2.7.0
torchvision==0.22.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.51.3
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
Werkzeug==3.1.3
